разбил на компоненты
